move ray path helpers from imaging to taup
